refactor: reorganize and rename block component modules for clarity
- Relocated block component files to `features/blocks` for improved structure.
- Renamed functions for better naming consistency (e.g., `load_glitch_component` to `loadGlitchBlockComponent`).
- Updated import paths across modules to reflect new organization.
