+++ behaviour_pack/scripts-dev/features/blocks/monolithic-reactor.ts
@@ -5,9 +5,19 @@ import {
     EquipmentSlot, Player,
     system
 } from "@minecraft/server";
-import utils from "../utils";
+import utils from "../../utils";
 
-export default function load_reactor_activate_component() {
+export default function loadMonolithicReactorComponent() {
+    const GLITCHES = [
+        utils.commands.noise_glitch,
+        utils.commands.vision_block_glitch,
+        utils.commands.vision_entity_glitch,
+        utils.commands.effect_glitch,
+        utils.commands.noise_glitch,
+        utils.commands.noise_glitch,
+        utils.commands.noise_glitch,
+    ]
+
     function on_interact(event: BlockComponentPlayerInteractEvent) {
         const mainhand = event.player?.getComponent(EntityComponentTypes.Equippable)?.getEquipment(EquipmentSlot.Mainhand)
         if (mainhand?.typeId == 'amethyst:glitch_core') {
@@ -19,19 +29,9 @@ export default function load_reactor_activate_component() {
                 event.player?.getComponent(EntityComponentTypes.Equippable)?.setEquipment(EquipmentSlot.Mainhand)
                 event.dimension.playSound('beacon.activate', event.block.center())
 
-                const glitches_type = [
-                    utils.commands.noise_glitch,
-                    utils.commands.vision_block_glitch,
-                    utils.commands.vision_entity_glitch,
-                    utils.commands.effect_glitch,
-                    utils.commands.noise_glitch,
-                    utils.commands.noise_glitch,
-                    utils.commands.noise_glitch,
-                ]
-
                 event.block.dimension.getPlayers()
                     .forEach((player: Player) => {
-                        glitches_type.forEach((glitch) => {
+                        GLITCHES.forEach((glitch) => {
                             glitch(player)
                         })
                     })
